Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,8 +2,8 @@
 int main() {
     int a,b,c;
     scanf("%d%d%d",&a,&b,&c);
-    int sum=a+b+c;
-    float avg=sum/3;
+    float sum=a+b+c;
+    float avg=sum/3.0;
     printf("Average: %.2f",avg);
     return 0;
 }
